Add config file and debugging option
A configuration file ("config.yml") has been added to the project with the option to toggle debugging messages. This feature will allow users to enable or disable config debugging messages through this config file, improving debugging flexibility. The AquaticCratesTesting.java file was updated to load and apply this configuration setting.