Add NMS support for version 1.21
The commit involves the addition of NMS (Netty Messaging Service) support for version 1.21. This involves creating and modifying xml files for version 1_21 and updating the AquaticCrates class, the plugin config files (pom and plugin.yml), and the parent project pom. The commit provides a smoother integration with the Minecraft version 1.21 by defining the required dependencies, repositories, plugins and build properties.